requirements.txt
azure-common==1.1.28
azure-core==1.24.1
azure-functions==1.11.2
azure-identity==1.10.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.4.0
azure-keyvault-keys==4.5.1
azure-keyvault-secrets==4.4.0
azure-storage-blob==12.12.0
black==22.6.0
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Mail==0.9.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaleido==0.2.1
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.23.0
oauthlib==3.2.0
pandas==1.4.3
pathspec==0.9.0
platformdirs==2.5.2
plotly==5.9.0
portalocker==2.4.0
protobuf==4.21.2
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.0.3
tenacity==8.0.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
